Move debounce function to own module
Status is using it, but autocomplete will be in a moment.
