+++ src/status.js
@@ -1,20 +1,6 @@
 import { createElement, Component } from 'preact' /** @jsx createElement */
+import { debounce } from './debounce'
 
-const debounce = function (func, wait, immediate) {
-  let timeout
-  return function () {
-    const context = this
-    const args = arguments
-    const later = function () {
-      timeout = null
-      if (!immediate) func.apply(context, args)
-    }
-    const callNow = immediate && !timeout
-    clearTimeout(timeout)
-    timeout = setTimeout(later, wait)
-    if (callNow) func.apply(context, args)
-  }
-}
 const statusDebounceMillis = 1400
 
 export default class Status extends Component {
